+++ migrations/UdapDb/Udap/Server/Migrations/UdapDb/UdapDbContextModelSnapshot.cs
@@ -412,17 +412,17 @@ namespace Udap.Server.Migrations.UdapDb
 
             modelBuilder.Entity("Udap.Server.Entities.Anchor", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<DateTime>("BeginDate")
                         .HasColumnType("datetime2");
 
-                    b.Property<long>("CommunityId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CommunityId")
+                        .HasColumnType("int");
 
                     b.Property<bool>("Enabled")
                         .HasColumnType("bit");
@@ -451,11 +451,11 @@ namespace Udap.Server.Migrations.UdapDb
 
             modelBuilder.Entity("Udap.Server.Entities.AnchorCertification", b =>
                 {
-                    b.Property<long>("AnchorId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("AnchorId")
+                        .HasColumnType("int");
 
-                    b.Property<long>("CertificationId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CertificationId")
+                        .HasColumnType("int");
 
                     b.HasKey("AnchorId", "CertificationId");
 
@@ -466,14 +466,14 @@ namespace Udap.Server.Migrations.UdapDb
 
             modelBuilder.Entity("Udap.Server.Entities.Certification", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<long?>("CommunityId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("CommunityId")
+                        .HasColumnType("int");
 
                     b.Property<string>("Name")
                         .IsRequired()
@@ -489,11 +489,11 @@ namespace Udap.Server.Migrations.UdapDb
 
             modelBuilder.Entity("Udap.Server.Entities.Community", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<bool>("Default")
                         .HasColumnType("bit");
@@ -513,11 +513,11 @@ namespace Udap.Server.Migrations.UdapDb
 
             modelBuilder.Entity("Udap.Server.Entities.CommunityCertification", b =>
                 {
-                    b.Property<long>("CommunityId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CommunityId")
+                        .HasColumnType("int");
 
-                    b.Property<long>("CertificationId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("CertificationId")
+                        .HasColumnType("int");
 
                     b.HasKey("CommunityId", "CertificationId");
 
@@ -528,11 +528,11 @@ namespace Udap.Server.Migrations.UdapDb
 
             modelBuilder.Entity("Udap.Server.Entities.RootCertificate", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<DateTime>("BeginDate")
                         .HasColumnType("datetime2");
